refactor: extract service manager runtime tests
